重构客户端底层数localDataSource,为其封装接口并统一命名, 重写底部导航栏UI

+++ shared/src/commonMain/kotlin/app/penny/core/domain/usecase/GetTransactionsCountUseCase.kt
@@ -6,7 +6,7 @@ class GetTransactionsCountUseCase(
     private val transactionRepository: TransactionRepository
 ) {
 
-    suspend operator fun invoke(): Int {
+    suspend operator fun invoke(): Long {
         val count = transactionRepository.count()
         return count
     }
